requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
numpy==1.19.4
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
